+++ bin/main.js
@@ -4,9 +4,10 @@ var nopt    = require('nopt')
   , url     = require('url')
   , PouchDB = require('pouchdb')
   , Spinner = require('cli-spinner').Spinner
+  , _       = require('lodash')
   , digest  = require('../lib')
-  , mapKeys = require('lodash/object/mapKeys')
   , merge   = require('../lib/merge')
+  , report  = require('../lib/report')
 
 PouchDB.plugin(require('pouchdb-migrate'))
 
@@ -14,10 +15,12 @@ var manifest = require('../package.json')
   , knownOptions = { 'help' : Boolean
                    , 'version' : Boolean
                    , 'allow-duplicates' : Boolean
+                   , 'report' : Boolean
                    , 'debug' : Boolean
                    }
   , shortHands   = { 'h' : ['--help']
                    , 'v' : ['--version']
+                   , 'r' : ['--report']
                    , 'A' : ['--allow-duplicates']
                    }
   , options      = nopt(knownOptions, shortHands)
@@ -34,6 +37,7 @@ function showHelp () {
   console.log('  -v, --version             print version')
   console.log('  -h, --help                show help message')
   console.log('  -A, --allow-duplicates    do not aggregate duplicates')
+  console.log('  -r, --report              print exhaustive report to stdout')
   console.log('  --debug                   print debug info')
   console.log('')
   console.log('Examples:')
@@ -58,6 +62,12 @@ function parseUrl (arg) {
   return arg.href
 }
 
+function mergedAndSources (doc) {
+  return { merged: 1
+         , sources: doc.sources && doc.sources.length
+         }
+}
+
 if (options.version) {
   printVersion()
   process.exit()
@@ -73,27 +83,74 @@ if (options.database) {
   var database   = new PouchDB(options.database)
     , spinner    = new Spinner(chalk.blue('%s Extracting documents...'))
     , aggregator = (options['allow-duplicates']) ? null : Object.create(null)
+    , reporter   = report()
 
+  if (options.report) {
+    reporter.table('merged docs', [ ['Case Id'    , 'case.id']
+                                  , ['Surname'    , 'surname']
+                                  , ['Other Names', 'otherNames']
+                                  , ['Phone'      , 'phoneNumber']
+                                  ]
+                  )
+  }
+
+  if (options.debug) {
+    reporter.table('merged sources', [ ['Case Id' , 'doc.id']
+                                     , ['Name'    , 'doc.name']
+                                     , ['Phone'   , 'doc.phone']
+                                     , ['Relative', 'doc.relative']
+                                  ]
+                  )
+  }
+
+  reporter.start()
   spinner.start()
+
   database
     .migrate(function (doc) {
-      return digest(doc, aggregator)
+      var results = digest(doc, aggregator)
+      reporter.count(results)
+      return results
     })
     .then(function (migration) {
       if (aggregator) {
-        var rels = mapKeys( aggregator
+        var rels = _.mapKeys( aggregator
                             , function (rel, key) { return rel[0]._id }
                             )
         spinner.stop(true)
         spinner = new Spinner(chalk.blue('%s Merging duplicates...'))
         spinner.start()
         return database.migrate(function (doc) {
-          return merge(doc, rels)
+          var results = merge(doc, rels)
+          reporter.count(results, mergedAndSources)
+          reporter.rows('merged docs', results)
+          if (options.debug) {
+            reporter.rows('merged sources', results && results[0].sources)
+          }
+          return results
         })
       }
     })
     .then(function (migration) {
       spinner.stop(true)
+      reporter.stop()
+      console.log(chalk.green('Done.'))
+      console.log(chalk.grey('Duration:', reporter.duration))
+      console.log( 'Updated:', reporter.updated
+                 , 'Created:', reporter.created
+                 , 'Total:'  , reporter.total
+                 )
+      var merged = reporter.amount('merged')
+      if (merged) {
+        console.log('Merged:', merged, 'from', reporter.amount('sources'))
+        console.log(reporter.render('merged docs'))
+        if (options.debug) {
+          console.log('')
+          console.log('===========================================================================')
+          console.log('')
+          console.log(reporter.render('merged sources'))
+        }
+      }
       process.exit()
     })
     .catch(function (error) {